scripts/requirements.txt
google-genai==1.21.1
pytesseract==0.3.13
opencv-python==4.11.0.86
numpy==2.3.1
pillow==11.2.1
pdf2image==1.17.0
pathlib==1.0.1
python-dotenv==1.1.1
regex==2024.11.6
